requirements.txt
pillow>=10.4
numpy>=1.26
